# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_controller.AnyMap:
    additionalProperties: true
    type: object
...
  internal_controller.K8sListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/internal_controller.AnyMap'
        type: array
    type: object
...
  internal_controller.K8sTextResp:
    properties:
      text:
        type: string
    type: object
...
  k8s-platform-backend_pkg_resp.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 星枢K8S管理平台后端接口文档（统一响应结构：code/message/data）
  title: 星枢K8S管理平台 API
  version: "1.0"
paths:
...
  /clusters/{id}/nodes:
    get:
      consumes:
      - application/json
      description: 获取指定集群节点列表
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        example: desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/internal_controller.K8sListResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 节点列表
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/cordon:
    post:
      consumes:
      - application/json
      description: 将节点标记为不可调度
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 停止节点调度
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/detail:
    get:
      consumes:
      - application/json
      description: 获取指定 Node 的详细信息（原始对象）
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 获取 Node 详情
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/events:
    get:
      consumes:
      - application/json
      description: 获取指定节点相关的 Event 列表
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/internal_controller.K8sListResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: Node Events 列表
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/pods:
    get:
      consumes:
      - application/json
      description: 获取指定节点上的 Pod 列表
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        example: desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/internal_controller.K8sListResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: Node Pods 列表
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/uncordon:
    post:
      consumes:
      - application/json
      description: 将节点标记为可调度
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 恢复节点调度
      tags:
      - K8s 资源接口
  /clusters/{id}/nodes/{name}/yaml:
    get:
      consumes:
      - application/json
      description: 获取指定 Node 的 YAML 文本
      parameters:
      - description: 集群ID
        example: 1001
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/internal_controller.K8sTextResp'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 获取 Node YAML
      tags:
      - K8s 资源接口
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"